Iteration H1:Sending Confirmation Emails and Iteration H2:Integration Testing of Applications
We configured our development, test, and production environments for our rails appliication to enable the sending of outbound emails. We created and tailored a mailer that'll send confirmation emails in both plain-text and HTML formats for people who ordered our products. We created both a functional test for the emails produces and integration test that covers the entire order scenario.